Add external privacy policy for surveys
https://eaflood.atlassian.net/browse/WATER-5037

To initiate surveys and user research, a privacy policy is in place and being used for WRWA research. This privacy policy is currently hosted on a 3rd part website (smart survey).

We want the privacy policy to be hosted by the service, to improve trust and confidence in its source.