Amend charter link in new research privacy policy (#2722)
https://eaflood.atlassian.net/browse/WATER-5037

To initiate surveys and user research, a privacy policy is in place and being used for WRWA research. This privacy policy is currently hosted on a third-party website (Smart Survey).

In [Add external privacy policy for surveys](https://github.com/DEFRA/water-abstraction-ui/pull/2721), we added the new page. But one of the links was slightly wrong.

This change fixes it.