Add snyk exception for tmp
https://github.com/DEFRA/water-abstraction-team/issues/144

**Snyk** has identified that [tmp](https://github.com/raszi/node-tmp) has a [Symlink Attack vulnerability](https://security.snyk.io/vuln/SNYK-JS-TMP-11501554).

Unfortunately, it is introduced via [@hapi/scooter](https://github.com/hapijs/scooter) which hasn't been updated in more than 3 years, and is unlikely too.

As this is an issue in the legacy code, the effort to find a replacement or resolve the issue puts it out of scope hence we are adding an exception.