Fix broken view licence summary returns test (#2731)
https://eaflood.atlassian.net/browse/WATER-4805

In [Fix view this return link on edit submit page](https://github.com/DEFRA/water-abstraction-ui/pull/2683), we added a feature flag for whether to use the legacy or new [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system) links to view a return.

However, _someone_ was naughty and didn't run the unit tests locally. If they had they would have spotted that the flag when set to `true` was causing a test to fail! 🤦😬

This change fixes the test by stubbing the config to set the feature flag to false.