Update filter used in /last/email endpoint
https://eaflood.atlassian.net/browse/WATER-5229

The legacy code always populates the `send_after` field with the current time when creating a notification, and only sometimes the `date_created` field.

The new code in [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system) always populates `date_created`, and never populates `send_after`. We now consider it a redundant field.

Our issue is that our acceptance tests rely on an endpoint that returns the last notification created for a specified email address. But the disparity between how the two systems create notifications means it can no longer find the last email, which has broken some of the tests.

We need to fix it so it can!

In [Auto-populate scheduled_notification date_created](https://github.com/DEFRA/water-abstraction-service/pull/2716) we added a migration to

- update the `water.scheduled_notification` table so that `date_created` will now automatically populate when a new record is inserted
- update existing records setting `date_created` to `send_after` where `date_created` is null and `send_after` isn't

Knowing that `date_created` will now always be populated for both legacy and system notifications we can use it rather than `send_after` in the filter `src/shared/lib/connectors/services/water/NotificationsApiClient.js` compiles when generating the request for the last email created.