Return search - Not due yet status https://eaflood.atlassian.net/browse/WATER-5216
Currently when a return log is created it is automatically pre-populated with a due date. However, the process of assigning a due date to a return log is being updated so that the due date is not set on creation. Instead, it is set when the reminders are sent out.

This means that some new return logs will now have a null due date where previously the due date would always have been populated.

This PR will update the search results so that when searching for a return. The status of a return will display ‘Not Due Yet’ when the due date is not set and the status is "due".