Fix error when viewing NALD gap charge versions
https://eaflood.atlassian.net/browse/WATER-5258

Charge versions created in WRLS do not allow gaps in the charge version history. When NALD managed charge versions, it did.

When WRLS took over charging from NALD, the previous team opted to fill these gaps with 'NALD GAP' charge versions. These are essentially empty records not linked to billing accounts.

When you view the licence summary's setup tab, these charge versions are listed with the others, and have a view link. However, if you click the link, you get an error.

The issue is that they have no billing account, but the controller is written with the assumption every charge version has one.

By tweaking the controller code, we can prevent an error, not that there is much to see!