Implement new OPEN return log status
https://eaflood.atlassian.net/browse/WATER-5288

Two things coincided recently. The first was WATER-5144, an issue reported by our users regarding the inability to submit return logs. The second is that we have started work on implementing 'dynamic due dates'.

Currently, when we create a return log, we apply a 'due date' 28 days after the end of the returns cycle it is linked to. For example, a return log linked to the winter & all-year cycle will be created as

- Start date 1 April 2025
- End date 31 March 2026
- Due date 28 April 2026

Confusion about this is what led to WATER-5144. Some users thought the due date was set as 28 days after the return log end date. This is easier to see on a 'split log'. For example, a licence revoked partway through the year will result in the current return log being voided and a new one created.

- Start date 1 April 2025
- End date 15 June 2025 (date licence revoked)
- Due date 28 April 2026

Those users would expect the due date to be 13 June 2025, and therefore, the return log could be submitted. But this not being the case, the return log will show as `NOT DUE YET` until 1 April 2026, when it then shows as `DUE`. At this point, it would also become available for submission.

Realising this, we were asked to make a change to the logic and [Allow ended 'Not due yet' returns to be submitted](https://github.com/DEFRA/water-abstraction-system/pull/2418).

So, now you can have two return logs, both showing as `NOT DUE YET`, one of which can be submitted and the other cannot. How does a user tell?

Adding to this confusion, we've started work on 'dynamic due dates'. Soon, when we create return logs, we will no longer assign a due date. Instead, it will be set when we receive confirmation from [GOV.UK Notify](https://www.notifications.service.gov.uk/) that an invitation to submit them has been sent. In the meantime, they too will display as `NOT DUE YET`. Again, the return log can be submitted once we are past its end date, but how would a user tell the difference?

The solution is a new `OPEN` status. This was tested with users in the early stages of implementing quarterly return logs and was positively received. However, that change wasn't necessary in the end. Now it is!

For clarity these are the statuses you'll see in the UI after this change.

- `NOT DUE YET` - The return log end date is greater than or equal to the current date. Cannot be submitted or received.
- `OPEN` - The return log end date is less than the current date, its DB status is 'due', and it either has a due date more than 28 days in the future, or it is null.
- `DUE` - The return log end date is less than the current date, its DB status is 'due', and it has a due date less than or equal to 28 days in the future.
- `OVERDUE` - The return log end date is less than the current date, its DB status is 'due', and it has a due date greater than the current date.
- `RECEIVED` - The return log has a DB status of `received`.
- `COMPLETE` - The return log has a DB status of `completed`.
- `VOID` - The return log has a DB status of `void`. Cannot be submitted or received.

We need to update the UI with matching logic to [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system) for those few times when a user searches for a return in the search page.