Remove enable system manage view feature flag
https://eaflood.atlassian.net/browse/WATER-5206

We recently created a new version of the 'Manage' page in [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system/pull/2283). As per normal, we also added a feature flag so we could control when we do the switch.

However, since then we have had to make two changes to the manage page (invalid addresses report and removing redundant notice links), and as we have a replacement there was no value in making the same changes to the legacy one.

Those changes need to be available to users, which makes the feature flag redundant: the system manage page is now the _only_ manage page.

This change removes the feature flag and sets the system one as the only version to use.