Promote to Notices nav bar (#2751)
https://eaflood.atlassian.net/browse/WATER-5347

We've done _a lot_ of work on notices and their view pages. The view notices page is now at a level that matches bill runs. It's a place where users can not only interrogate the data but also create new notices.

The end goal is to drop the links in the manage page completely and use the notices page.

Moving **Notices** into the top-level nav bar, rather than just as a link in the reports section, will indicate it is now a feature, not just a report.