Stop filtering non-current return logs externally
https://eaflood.atlassian.net/browse/WATER-5373

When the Billing & Data team started checking the recipients that would be selected for the next round of return invitations, they spotted that there were some missing (when compared to a report they ran).

When they looked, they were all linked to licences that had return versions with a reason of 'Succession or transfer of licence'.

When this is the reason against the return version, the return logs we generate from it have the flag `isCurrent`, found in their `metadata` field, set to 'false'.

We don't know why; it was something the legacy code did and we copied into [https://github.com/DEFRA/water-abstraction-system](water-abstraction-system)!

Another thing we copied was the legacy code's query for selecting which return logs to base a returns invitation on. The query excluded return logs where `isCurrent = false`, so we did the same. But after Billing & Data declared them missing, we [dropped the filter from the returns notice queries](https://github.com/DEFRA/water-abstraction-system/pull/2554).

The first summer returns invitation has just been sent with these `isCurrent=false` return logs now being included, which means additional licensees will receive a notification.

But this has led to the first report from one of them saying they can see nothing to submit when they log in. This is because as well as filtering them out of the returns invitation, the legacy code filters them out of the external view. So, now we've been asked to drop the filter there as well.