Update the default view width for the internal service
https://eaflood.atlassian.net/browse/WATER-5217

We need to increase the width of the internal service from the default '960px' set by the design syetem.

This change updated and the default 'govuk-width-container' class to ''max-width: 1100px !important;'

We use the !important flag to give this rule the highest priority, ensuring it overrides any other conflicting max-width declarations for .govuk-width-container.