Add feature flag for new search page
https://eaflood.atlassian.net/browse/WATER-5292

We are migrating the internal search page (the root page of the internal service) from the legacy apps to our new `water-abstraction-system` app.

The new search page has been created, this change is to add a new feature flag `USE_SYSTEM_SEARCH` that tells the application to use the new search page instead of the old one.

To avoid multiple redirects, any links that point to the  `/licences' path have been updated to point to the correct root path `/` instead