Fix broken link to system return logs (#2775)
https://eaflood.atlassian.net/browse/WATER-5353

In [Use new return log id](https://github.com/DEFRA/water-abstraction-ui/pull/2752), we updated the UI to use the new `ID` property we'd added to the return logs table, which is a UUID.

However, we've spotted a page where that change broke the link. We need to ensure that, in this case, the helper includes the field when fetching the data. This change means the value is then available to assign to the link.